backend/requirements.txt
Flask
Flask-Cors
pdfplumber
groq
python-dotenv
